prueba15

+++ src/main/java/com/reniec/ren/repository/ReclamoRepository.java
@@ -3,5 +3,5 @@ package com.reniec.ren.repository;
 import com.reniec.ren.modelo.Reclamo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReclamoRepository extends JpaRepository<Reclamo, Long> {
+public interface ReclamoRepository extends JpaRepository<Reclamo, Integer> {
 }
